analyzer: stubs for serializing types.
Also move type_type to .cpp file.
