+++ include/reaver/vapor/analyzer/types/function.h
@@ -1,7 +1,7 @@
 /**
  * Vapor Compiler Licence
  *
- * Copyright © 2017 Michał "Griwes" Dominiak
+ * Copyright © 2017-2018 Michał "Griwes" Dominiak
  *
  * This software is provided 'as-is', without any express or implied
  * warranty. In no event will the authors be held liable for any damages
@@ -48,6 +48,11 @@ inline namespace _v1
         }
 
         virtual future<std::vector<function *>> get_candidates(lexer::token_type op) const override;
+
+        virtual std::unique_ptr<proto::type> generate_interface() const override
+        {
+            assert(0);
+        }
 
     private:
         virtual void _codegen_type(ir_generation_context & ctx) const override
