+++ include/reaver/vapor/analyzer/types/pack.h
@@ -1,7 +1,7 @@
 /**
  * Vapor Compiler Licence
  *
- * Copyright © 2017 Michał "Griwes" Dominiak
+ * Copyright © 2017-2018 Michał "Griwes" Dominiak
  *
  * This software is provided 'as-is', without any express or implied
  * warranty. In no event will the authors be held liable for any damages
@@ -59,6 +59,11 @@ inline namespace _v1
         virtual type * get_pack_type() const override
         {
             assert(!"tried to make a pack of packs, this is nonsense");
+        }
+
+        virtual std::unique_ptr<proto::type> generate_interface() const override
+        {
+            assert(0);
         }
 
     private:
