+++ include/reaver/vapor/analyzer/types/sized_integer.h
@@ -1,7 +1,7 @@
 /**
  * Vapor Compiler Licence
  *
- * Copyright © 2017 Michał "Griwes" Dominiak
+ * Copyright © 2017-2018 Michał "Griwes" Dominiak
  *
  * This software is provided 'as-is', without any express or implied
  * warranty. In no event will the authors be held liable for any damages
@@ -102,6 +102,11 @@ inline namespace _v1
         auto size() const
         {
             return _size;
+        }
+
+        virtual std::unique_ptr<proto::type> generate_interface() const override
+        {
+            assert(0);
         }
 
     private:
