+++ lib/proto/entity.proto
@@ -26,7 +26,10 @@ package reaver.vapor.proto;
 
 import "google/protobuf/any.proto";
 
+import "type.proto";
+
 message entity
 {
+    type type = 1;
     google.protobuf.Any value = 2;
 }
